useMemo for mutation result
I had a use case where I wanted to run some effect only if the mutation result was different. 
But just putting the mutation result as a dependency of the useEffect didn't work; If the mutation was fired again, even if the result was identical to the first one, the effect ran again because the result was a new value.
With this fix, the result is only created again if it was also created again from the js side.